requirements.txt
fastmcp
httpx
python-dotenv
pydantic>=2,<3
python-docx
